+++ src/app/events/page.tsx
@@ -1,11 +1,11 @@
 import Filters from "@/components/Filters";
-import Map from "@/components/Map";
+import MapContainer from "@/components/MapContainer";
 import React from "react";
 
 const EventsPage = () => {
   return (
     <div className='grid grid-cols-[3fr_1fr] w-full p-4'>
-      <Map />
+      <MapContainer />
       <Filters />
     </div>
   );
